src/955B/test_cdf_955B.py
import unittest
from unittest.mock import patch

from cdf_955B import CodeforcesTask955BSolution


class TestCDF955B(unittest.TestCase):
    def test_955B_acceptance_1(self):
        mock_input = ['ababa']
        expected = 'Yes'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask955BSolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)

    def test_955B_acceptance_2(self):
        mock_input = ['zzcxx']
        expected = 'Yes'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask955BSolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)

    def test_955B_acceptance_3(self):
        mock_input = ['yeee']
        expected = 'No'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask955BSolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)


if __name__ == "__main__":
    unittest.main()
